Polish Kafka meter binder changes
